gh-943 add class search (without vector)
to be used in traverser.Get
